streaming_jobs/kafka_python_demo/src/stratus_streaming/requirements.txt
kafka-python==2.0.2

